Start of the series
Pushing changes that matter onto clean branch known as "originalForm"

Pushing changes that matter onto clean branch known as "originalForm"

Pushing changes that matter onto clean branch known as "originalForm"

Pushing changes that matter onto clean branch known as "originalForm"

Hmm.

Hmm.

save compiledload and mp3obj

Conflicts:
	geode-core/src/main/java/org/apache/geode/cache/query/internal/CompiledLoad.java
	geode-core/src/main/java/org/apache/geode/cache/query/internal/DefaultQuery.java
	geode-core/src/main/java/org/apache/geode/cache/query/internal/Mp3Obj.java
	geode-core/src/main/java/org/apache/geode/cache/query/internal/QCompiler.java
	geode-core/src/main/java/org/apache/geode/cache/query/internal/parse/ASTLoad.java
	geode-core/src/main/java/org/apache/geode/cache/query/internal/parse/GemFireAST.java
	geode-core/src/main/java/org/apache/geode/cache/query/internal/parse/oql.g

Hmm.

Mp3 constructor has been made more proper....and CompiledLoader is now going to put stuff onto a region.

Slight tweak to CompiledLoad

Slight tweak to CompiledLoad

Slight tweak to CompiledLoad

Slight tweak to CompiledLoad

Slight tweak to CompiledLoad

Slight tweak to CompiledLoad

Slight tweak to CompiledLoad

Slight tweak to CompiledLoad

Slight tweak to CompiledLoad

Slight tweak to CompiledLoad

Slight tweak to CompiledLoad

Slight tweak to CompiledLoad

Slight tweak to CompiledLoad

Involved the use of using a className

Involved the use of using a className

Involved the use of using a className

Involved the use of using a className

Involved the use of using a className

Involved the use of using a className

Involved the use of using a className

Involved the use of using a className

Involved the use of using a className

Involved the use of using a className

Involved the use of using a className

Involved the use of using a className

Involved the use of using a className

Involved the use of using a className

Involved the use of using a className

Involved the use of using a className

Involved the use of using a className

Involved the use of using a className

Involved the use of using a className

Involved the use of using a className

Involved the use of using a className

Involved the use of using a className

Involved the use of using a className

Involved the use of using a className

Involved the use of using a className

Involved the use of using a className

Involved the use of using a className

Involved the use of using a className

Involved the use of using a className

Involved the use of using a className

Involved the use of using a className

Involved the use of using a className

Involved the use of using a className

Involved the use of using a className

Involved the use of using a className

Involved the use of using a className

Involved the use of using a className

Involved the use of using a className

Involved the use of using a className

Involved the use of using a className

Involved the use of using a className

Involved the use of using a className

Involved the use of using a className